refactor: GamePieceExceptPawn 폰을 제외한 기물을 추상 클래스로 묶어 중복로직 제거
