+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,12 +1,12 @@
 package chess.domain.piece;
 
 import chess.domain.game.Side;
+import chess.domain.gamestate.GamePieceExceptPawn;
 import chess.domain.position.Position;
-import chess.exception.InvalidMethodCallException;
 import java.util.Collections;
 import java.util.List;
 
-public final class Knight extends GamePiece {
+public final class Knight extends GamePieceExceptPawn {
 
     private static final String INITIAL = "N";
     private static final double SCORE = 2.5;
@@ -33,11 +33,6 @@ public final class Knight extends GamePiece {
     }
 
     @Override
-    public boolean isPawn() {
-        return false;
-    }
-
-    @Override
     public boolean isKing() {
         return false;
     }
@@ -45,15 +40,5 @@ public final class Knight extends GamePiece {
     @Override
     public double score() {
         return SCORE;
-    }
-
-    @Override
-    public boolean diagonal(Position from, Position to) {
-        throw new InvalidMethodCallException();
-    }
-
-    @Override
-    public boolean forward(Position from, Position to) {
-        throw new InvalidMethodCallException();
     }
 }
