+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,11 +1,11 @@
 package chess.domain.piece;
 
 import chess.domain.game.Side;
+import chess.domain.gamestate.GamePieceExceptPawn;
 import chess.domain.position.Position;
-import chess.exception.InvalidMethodCallException;
 import java.util.List;
 
-public final class Queen extends GamePiece {
+public final class Queen extends GamePieceExceptPawn {
 
     private static final String INITIAL = "Q";
     private static final int SCORE = 9;
@@ -36,11 +36,6 @@ public final class Queen extends GamePiece {
     }
 
     @Override
-    public boolean isPawn() {
-        return false;
-    }
-
-    @Override
     public boolean isKing() {
         return false;
     }
@@ -48,15 +43,5 @@ public final class Queen extends GamePiece {
     @Override
     public double score() {
         return SCORE;
-    }
-
-    @Override
-    public boolean diagonal(Position from, Position to) {
-        throw new InvalidMethodCallException();
-    }
-
-    @Override
-    public boolean forward(Position from, Position to) {
-        throw new InvalidMethodCallException();
     }
 }
